requirements.txt
streamlit
feedparser
google-generativeai
pandas
beautifulsoup4
python-dotenv
lxml
